# docs/portfolio_analysis_api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FiMatchPlus Portfolio Analysis API",
    "description": "이동 윈도우 기반 MPT 포트폴리오 최적화 및 백테스팅 분석 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/analysis/start": {
      "post": {
        "summary": "비동기 포트폴리오 분석 시작",
        "description": "포트폴리오 분석을 백그라운드에서 실행하고 완료 시 콜백 URL로 결과를 전송합니다. 이동 윈도우 기반 MPT 최적화 및 백테스팅 분석을 수행합니다.",
        "tags": [
          "analysis"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              },
              "example": {
                "holdings": [
                  {
                    "stock_code": "005930",
                    "quantity": 100
                  },
                  {
                    "stock_code": "000660",
                    "quantity": 50
                  },
                  {
                    "stock_code": "035420",
                    "quantity": 30
                  }
                ],
                "lookback_years": 5,
                "benchmark": "KOSPI",
                "risk_free_rate": 0.035,
                "callback_url": "https://your-server.com/analysis/callback",
                "analysis_id": 12345
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "분석 시작 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJobResponse"
                },
                "example": {
                  "job_id": "550e8400-e29b-41d4-a716-446655440000",
                  "status": "started",
                  "message": "포트폴리오 분석이 백그라운드에서 실행 중입니다."
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "callback_url is required for async analysis",
                  "detail": "비동기 분석을 위해서는 콜백 URL이 필요합니다.",
                  "timestamp": "2024-01-01T10:00:00Z"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Internal server error",
                  "detail": "분석 시작 중 오류가 발생했습니다.",
                  "timestamp": "2024-01-01T10:00:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/run": {
      "post": {
        "summary": "동기 포트폴리오 분석 실행",
        "description": "포트폴리오 분석을 즉시 실행하고 결과를 반환합니다. 3년 이동 윈도우 기반 MPT 최적화 및 백테스팅 분석을 수행합니다. 최근 시점의 최적 비중과 백테스팅 기반 검증된 성능 지표를 제공합니다.",
        "tags": [
          "analysis"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              },
              "example": {
                "holdings": [
                  {
                    "stock_code": "005930",
                    "quantity": 100
                  },
                  {
                    "stock_code": "000660",
                    "quantity": 50
                  },
                  {
                    "stock_code": "035420",
                    "quantity": 30
                  }
                ],
                "lookback_years": 5,
                "benchmark": "KOSPI",
                "risk_free_rate": 0.035
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "분석 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnhancedAnalysisResponse"
                },
                "example": {
                  "success": true,
                  "min_variance": {
                    "weights": {
                      "005930": 0.45,
                      "000660": 0.35,
                      "035420": 0.20
                    }
                  },
                  "max_sharpe": {
                    "weights": {
                      "005930": 0.50,
                      "000660": 0.30,
                      "035420": 0.20
                    }
                  },
                  "metrics": {
                    "min_variance": {
                      "expected_return": 0.085,
                      "std_deviation": 0.15,
                      "beta": 0.95,
                      "alpha": 0.012,
                      "jensen_alpha": 0.008,
                      "tracking_error": 0.08,
                      "sharpe_ratio": 0.33,
                      "treynor_ratio": 0.053,
                      "sortino_ratio": 0.42,
                      "calmar_ratio": 0.28,
                      "information_ratio": 0.15,
                      "max_drawdown": -0.22,
                      "downside_deviation": 0.12,
                      "var_value": -0.15,
                      "cvar_value": -0.19,
                      "upside_beta": 1.05,
                      "downside_beta": 0.85,
                      "correlation_with_benchmark": 0.78
                    },
                    "max_sharpe": {
                      "expected_return": 0.092,
                      "std_deviation": 0.18,
                      "beta": 1.02,
                      "alpha": 0.018,
                      "jensen_alpha": 0.014,
                      "tracking_error": 0.10,
                      "sharpe_ratio": 0.32,
                      "treynor_ratio": 0.056,
                      "sortino_ratio": 0.48,
                      "calmar_ratio": 0.32,
                      "information_ratio": 0.18,
                      "max_drawdown": -0.25,
                      "downside_deviation": 0.14,
                      "var_value": -0.18,
                      "cvar_value": -0.23,
                      "upside_beta": 1.08,
                      "downside_beta": 0.92,
                      "correlation_with_benchmark": 0.82
                    }
                  },
                  "benchmark_comparison": {
                    "benchmark_code": "KOSPI",
                    "benchmark_return": 0.065,
                    "benchmark_volatility": 0.16,
                    "excess_return": 0.027,
                    "relative_volatility": 1.125,
                    "security_selection": 0.019,
                    "timing_effect": 0.008
                  },
                  "risk_free_rate_used": 0.035,
                  "analysis_period": {
                    "start": "2019-01-01T00:00:00Z",
                    "end": "2024-01-01T00:00:00Z"
                  },
                  "notes": "Analysis based on KOSPI benchmark and 3-year rolling window optimization."
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "At least one holding must be specified",
                  "detail": "포트폴리오에 최소 1개 이상의 종목이 필요합니다.",
                  "timestamp": "2024-01-01T10:00:00Z"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Internal server error",
                  "detail": "분석 중 오류가 발생했습니다.",
                  "timestamp": "2024-01-01T10:00:00Z"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "type": "object",
        "required": [
          "holdings"
        ],
        "properties": {
          "holdings": {
            "type": "array",
            "description": "보유 종목 목록 (최소 1개 이상 필요)",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/AnalysisHoldingInput"
            }
          },
          "lookback_years": {
            "type": "integer",
            "description": "과거 데이터 조회 연수 (3~5 추천, 최소 5년 권장)",
            "minimum": 1,
            "maximum": 10,
            "default": 3
          },
          "benchmark": {
            "type": "string",
            "description": "벤치마크 지수 코드 (예: KOSPI, KOSDAQ). 미제공 시 KOSPI 자동 사용",
            "example": "KOSPI"
          },
          "risk_free_rate": {
            "type": "number",
            "description": "연간 무위험 수익률 (소수). 미제공 시 자동 조회",
            "minimum": 0,
            "maximum": 1,
            "example": 0.035
          },
          "callback_url": {
            "type": "string",
            "description": "결과를 받을 콜백 URL (비동기 처리 시 필수)",
            "example": "https://your-server.com/analysis/callback"
          },
          "analysis_id": {
            "type": "integer",
            "description": "클라이언트에서 제공하는 분석 ID (콜백 시 그대로 반환)",
            "example": 12345
          }
        }
      },
      "AnalysisHoldingInput": {
        "type": "object",
        "required": [
          "stock_code",
          "quantity"
        ],
        "properties": {
          "stock_code": {
            "type": "string",
            "description": "종목코드 (6자리 숫자)",
            "pattern": "^[0-9]{6}$",
            "example": "005930"
          },
          "quantity": {
            "type": "integer",
            "description": "보유 수량 (주)",
            "minimum": 1,
            "example": 100
          }
        }
      },
      "EnhancedAnalysisResponse": {
        "type": "object",
        "required": [
          "success",
          "min_variance",
          "max_sharpe",
          "metrics",
          "risk_free_rate_used",
          "analysis_period"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "분석 성공 여부"
          },
          "min_variance": {
            "$ref": "#/components/schemas/PortfolioWeights",
            "description": "최소 분산 포트폴리오 비중 (최근 시점 기준)"
          },
          "max_sharpe": {
            "$ref": "#/components/schemas/PortfolioWeights",
            "description": "최대 샤프 비율 포트폴리오 비중 (최근 시점 기준)"
          },
          "metrics": {
            "type": "object",
            "description": "백테스팅 기반 평균 성능 지표",
            "properties": {
              "min_variance": {
                "$ref": "#/components/schemas/EnhancedAnalysisMetrics",
                "description": "최소 분산 포트폴리오의 백테스팅 성능 지표"
              },
              "max_sharpe": {
                "$ref": "#/components/schemas/EnhancedAnalysisMetrics",
                "description": "최대 샤프 포트폴리오의 백테스팅 성능 지표"
              }
            }
          },
          "benchmark_comparison": {
            "$ref": "#/components/schemas/BenchmarkComparison",
            "description": "벤치마크 비교 분석 결과 (벤치마크 제공 시)"
          },
          "risk_free_rate_used": {
            "type": "number",
            "description": "분석에 사용된 무위험수익률",
            "example": 0.035
          },
          "analysis_period": {
            "type": "object",
            "description": "분석 기간 (백테스팅 전체 기간)",
            "properties": {
              "start": {
                "type": "string",
                "format": "date-time",
                "description": "분석 시작일"
              },
              "end": {
                "type": "string",
                "format": "date-time",
                "description": "분석 종료일"
              }
            }
          },
          "notes": {
            "type": "string",
            "description": "참고 사항",
            "example": "Analysis based on KOSPI benchmark and 3-year rolling window optimization."
          },
          "execution_time": {
            "type": "number",
            "description": "실행 시간 (초)",
            "example": 2.5
          },
          "analysis_id": {
            "type": "integer",
            "description": "클라이언트에서 제공한 분석 ID",
            "example": 12345
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "응답 생성 시각",
            "example": "2024-01-01T10:00:00Z"
          }
        }
      },
      "AnalysisJobResponse": {
        "type": "object",
        "required": [
          "job_id",
          "status",
          "message"
        ],
        "properties": {
          "job_id": {
            "type": "string",
            "description": "작업 ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "status": {
            "type": "string",
            "description": "작업 상태",
            "example": "started"
          },
          "message": {
            "type": "string",
            "description": "상태 메시지",
            "example": "포트폴리오 분석이 백그라운드에서 실행 중입니다."
          }
        }
      },
...
      "PortfolioWeights": {
        "type": "object",
        "required": [
          "weights"
        ],
        "properties": {
          "weights": {
            "type": "object",
            "description": "종목코드별 비중 (합계 1.0)",
            "additionalProperties": {
              "type": "number",
              "minimum": 0,
              "maximum": 1
            },
            "example": {
              "005930": 0.45,
              "000660": 0.35,
              "035420": 0.20
            }
          }
        }
      },
      "EnhancedAnalysisMetrics": {
        "type": "object",
        "description": "향상된 포트폴리오 성과 지표 (백테스팅 기반 평균)",
        "required": [
          "expected_return",
          "std_deviation",
          "beta",
          "alpha",
          "jensen_alpha",
          "tracking_error",
          "sharpe_ratio",
          "treynor_ratio",
          "sortino_ratio",
          "calmar_ratio",
          "information_ratio",
          "max_drawdown",
          "downside_deviation",
          "upside_beta",
          "downside_beta",
          "correlation_with_benchmark"
        ],
        "properties": {
          "expected_return": {
            "type": "number",
            "description": "기대수익률 (연환산)",
            "example": 0.085
          },
          "std_deviation": {
            "type": "number",
            "description": "표준편차 (연환산)",
            "example": 0.15
          },
          "beta": {
            "type": "number",
            "description": "베타 (벤치마크 대비)",
            "example": 0.95
          },
          "alpha": {
            "type": "number",
            "description": "알파 (CAPM 기준)",
            "example": 0.012
          },
          "jensen_alpha": {
            "type": "number",
            "description": "젠센 알파",
            "example": 0.008
          },
          "tracking_error": {
            "type": "number",
            "description": "트래킹 에러",
            "example": 0.08
          },
          "sharpe_ratio": {
            "type": "number",
            "description": "샤프 비율",
            "example": 0.33
          },
          "treynor_ratio": {
            "type": "number",
            "description": "트레이너 비율",
            "example": 0.053
          },
          "sortino_ratio": {
            "type": "number",
            "description": "소르티노 비율",
            "example": 0.42
          },
          "calmar_ratio": {
            "type": "number",
            "description": "칼마 비율",
            "example": 0.28
          },
          "information_ratio": {
            "type": "number",
            "description": "정보비율",
            "example": 0.15
          },
          "max_drawdown": {
            "type": "number",
            "description": "최대 낙폭",
            "example": -0.22
          },
          "downside_deviation": {
            "type": "number",
            "description": "하방편차",
            "example": 0.12
          },
          "var_value": {
            "type": "number",
            "description": "VaR 95% (Value at Risk)",
            "example": -0.15
          },
          "cvar_value": {
            "type": "number",
            "description": "CVaR 95% (Conditional Value at Risk)",
            "example": -0.20
          },
          "upside_beta": {
            "type": "number",
            "description": "상승 베타",
            "example": 1.05
          },
          "downside_beta": {
            "type": "number",
            "description": "하락 베타",
            "example": 0.85
          },
          "correlation_with_benchmark": {
            "type": "number",
            "description": "벤치마크와의 상관관계",
            "example": 0.78
          }
        }
      },
      "BenchmarkComparison": {
        "type": "object",
        "description": "벤치마크 비교 분석 결과",
        "required": [
          "benchmark_code",
          "benchmark_return",
          "benchmark_volatility",
          "excess_return",
          "relative_volatility",
          "security_selection",
          "timing_effect"
        ],
        "properties": {
          "benchmark_code": {
            "type": "string",
            "description": "벤치마크 지수 코드",
            "example": "KOSPI"
          },
          "benchmark_return": {
            "type": "number",
            "description": "벤치마크 수익률 (연환산)",
            "example": 0.065
          },
          "benchmark_volatility": {
            "type": "number",
            "description": "벤치마크 변동성",
            "example": 0.16
          },
          "excess_return": {
            "type": "number",
            "description": "초과 수익률",
            "example": 0.027
          },
          "relative_volatility": {
            "type": "number",
            "description": "상대 변동성",
            "example": 1.125
          },
          "security_selection": {
            "type": "number",
            "description": "종목선택 효과",
            "example": 0.019
          },
          "timing_effect": {
            "type": "number",
            "description": "타이밍 효과",
            "example": 0.008
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "timestamp"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "에러 메시지"
          },
          "detail": {
            "type": "string",
            "description": "상세 에러 정보"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "에러 발생 시간"
          }
        }
      }
    }
  },
...
}